move fire sky to r_sky.c, add R_UpdateSky function
