+++ src/r_sky.h
@@ -39,12 +39,19 @@ extern boolean linearsky;
 extern int skytexture;
 extern int skytexturemid;
 
-extern sky_t *skydef;
+extern sky_t *sky;
 
 // Called whenever the view size changes.
 void R_InitSkyMap(void);
 
 byte R_GetSkyColor(int texturenum);
+
+void R_UpdateSky(void);
+
+#define FIRE_WIDTH     128
+#define FIRE_HEIGHT    256
+
+byte *R_GetFireColumn(int col);
 
 #endif
